client/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portscanner-client"
version = "1.0.0"
description = "Port Scanner Delta Reporter - Client Agent"
readme = "README.md"
authors = [{name = "Jared Stanbrook", email = "jared@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-nmap==0.7.1",
    "requests==2.31.0",
    "websockets==11.0.3",
    "Flask==2.3.3",
    "psutil==5.9.5",
    "PyYAML==6.0.1",
    "click==8.1.7",
    "questionary>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.2",
    "black==23.7.0",
    "flake8==6.0.0",
]

[project.scripts]
portscanner-client = "client_agent:main"
portscanner-client-config = "client_config:configure"

[tool.setuptools]
py-modules = ["client_agent", "client_config"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml"]